requirements.txt
algoliasearch==2.6.2
algoliasearch-django==2.0.0
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
decouple==0.0.7
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-environ-2==2.3.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
